requirements.txt
--prefer-binary
pip>=23.0
setuptools>=65.0
wheel


absl-py==1.4.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.3
cycler==0.10.0
cloudinary

Django==4.1.7
django-storages
django-environ
django-cloudinary-storage

dj-database-url 
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dotenv

et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
flatbuffers==23.5.26
fqdn==1.5.1
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
html5lib==1.1
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.13.1
kiwisolver==1.3.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.4.3
matplotlib-inline==0.1.6
mistune==3.0.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==3.8.0

prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg
psycopg2-binary
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyee==9.0.4
Pygments==2.15.1
pyparsing==2.4.7
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7

pytz==2023.3


PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2

scipy==1.10.1
seaborn==0.12.2
selenium==4.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1



tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0

tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.5.0
tzdata==2022.7
uri-template==1.3.0
urllib3==1.26.14
waitress>=2.1.2

wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
whitenoise
widgetsnbextension==4.0.7
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
